Move Headings to TS
