+++ src/gatsby-theme-docz/components/Headings/index.ts
@@ -1,0 +1,1 @@
+export { h1, h2, h3, h4, h5, h6 } from "./Headings";
